Move exceptions, and las_items to new modules
