+++ lasio/__init__.py
@@ -1,4 +1,6 @@
-from .las import __version__, read, LASFile, CurveItem, HeaderItem, JSONEncoder
+from las import (
+    read, LASFile, CurveItem, HeaderItem, JSONEncoder
+    )
 
 try:
     import openpyxl
@@ -6,3 +8,5 @@ except ImportError:
     pass
 else:
     from .excel import ExcelConverter
+
+__version__ = '0.10'
